Fixed bugs and added profanity and rich text

+++ src/components/Comment.js
@@ -1,10 +1,23 @@
-import { useState, useRef, useEffect } from "react";
+import IconButton from '@mui/material/IconButton';
+import Menu from '@mui/material/Menu';
+import MenuItem from '@mui/material/MenuItem';
+import MoreVertIcon from '@mui/icons-material/MoreVert';
+import * as React from 'react';
+import { useState, useRef, useEffect} from "react";
 import Action from "./Action";
 import { ReactComponent as DownArrow } from "../assets/down-arrow.svg";
 import { ReactComponent as UpArrow } from "../assets/up-arrow.svg";
 import { ReactComponent as Like } from "../assets/like.svg";
 import { ReactComponent as Dislike } from "../assets/dislike.svg";
-import axios from "axios";
+import { Editor } from 'react-draft-wysiwyg';
+import { EditorState, convertToRaw } from "draft-js";
+import draftToHtml from "draftjs-to-html";
+import { convertFromHTML } from "draft-convert";
+import 'react-draft-wysiwyg/dist/react-draft-wysiwyg.css';
+
+const profanity = require('leo-profanity');
+
+const ITEM_HEIGHT = 48;
 
 const Comment = ({
   handleInsertNode,
@@ -16,68 +29,115 @@ const Comment = ({
   email,
   comment,
 }) => {
-  const [input, setInput] = useState("");
+  const [editorState, setEditorState] = useState(EditorState.createEmpty());
+  const months = [
+    "January",
+    "February",
+    "March",
+    "April",
+    "May",
+    "June",
+    "July",
+    "August",
+    "September",
+    "October",
+    "November",
+    "December"
+  ];
+
+  const handleReset = () => {
+    setEditorState(EditorState.createEmpty());
+  };
+
   const [editMode, setEditMode] = useState(false);
   const [showInput, setShowInput] = useState(false);
   const [expand, setExpand] = useState(false);
   const inputRef = useRef(null);
+  const [editCommentState, setEditCommentState] = useState();
 
 
   useEffect(() => {
     inputRef?.current?.focus();
-    setExpand(comment.id === 1 | comment?.items?.length);
-    setInput("");
-
+    if (comment.id === 1)
+      setExpand(true)
+    else setExpand(false)
+    setEditorState(EditorState.createEmpty());
+    if (comment.id !== 1)
+      setEditCommentState(EditorState.createWithContent(convertFromHTML(comment?.name)))
   }, [editMode]);
 
   const handleNewComment = () => {
-    setExpand(!expand);
     setShowInput(true);
   };
 
+  const handleShowReplies = () => {
+    setExpand(!expand);
+  };
+
   const onAddComment = () => {
-    if(user === "")
-    {
+    if (user === "") {
       alert("Please Login to comment")
       return;
     }
     if (editMode) {
-      if(inputRef?.current?.innerText === "")
-      {
+      let htmlData = draftToHtml(convertToRaw(editCommentState.getCurrentContent()));
+      if (editCommentState.getCurrentContent().hasText() === false) {
         alert("Please type something");
         return;
       }
-      handleEditNode(comment.id, inputRef?.current?.innerText);
+      handleEditNode(comment.id, profanity.clean(htmlData));
     } else {
-      if (input === "") {
+      let htmlData = draftToHtml(convertToRaw(editorState.getCurrentContent()));
+      if (editorState.getCurrentContent().hasText() === false) {
         alert("Please type something");
         return;
       }
       setExpand(true);
-      handleInsertNode(comment.id, input);
+      console.log(htmlData)
+      console.log(profanity.clean(htmlData))
+      handleInsertNode(comment.id, profanity.clean(htmlData));
       setShowInput(false);
-      setInput("");
+      handleReset();
     }
 
     if (editMode) setEditMode(false);
   };
 
   const handleDelete = () => {
-    handleDeleteNode(comment.id);
+    handleClose();
+    if (window.confirm("Are you sure you want to delete") === true) {
+      handleDeleteNode(comment.id);
+    }
   };
+
+  const [anchorEl, setAnchorEl] = React.useState(null);
+  const open = Boolean(anchorEl);
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget);
+  };
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  const handleEdit = () => {
+    setEditMode(true)
+    handleClose();
+  };
+
+  var parse = require('html-react-parser');
+
+
 
   return (
     <div>
       <div className={comment.id === 1 ? "inputContainer" : "commentContainer"} >
         {comment.id === 1 ? (
           <>
-            <textarea
-              type="text"
-              className="inputContainer__input first_input"
-              autoFocus
-              value={input}
-              onChange={(e) => setInput(e.target.value)}
-              placeholder="type..."
+            <Editor
+              editorStyle={{ border: "1px solid #C0C0C0" }}
+              editorState={editorState}
+              onEditorStateChange={setEditorState}
+              placeholder="Type your comment here..."
             />
             <div className="comment-btn-container">
               <Action
@@ -89,16 +149,69 @@ const Comment = ({
           </>
         ) : (
           <>
-            <b>{comment.authorName}</b>
-            <br></br>
-            <span
-              contentEditable={editMode}
-              suppressContentEditableWarning={editMode}
-              ref={inputRef}
-              style={{ wordWrap: "break-word" }}
-            >
-              {comment.name}
-            </span>
+            <div style={{ display: "flex", marginTop: "5px" }}>
+              <p>
+                <b>{comment.authorName}</b>
+                <div className='lastUpdated'>
+                  Last updated at
+                  {
+                    // if new Date(comment.last_updated).getHours() is 1 digit, add 0 before it
+                    " " + (new Date(comment.last_updated).getHours().toString().length === 1 ? "0" : "") + new Date(comment.last_updated).getHours() + ":" + (new Date(comment.last_updated).getMinutes().toString().length === 1 ? "0": "") + new Date(comment.last_updated).getMinutes() + " " + new Date(comment.last_updated).getDate() + " " + months[new Date(comment.last_updated).getMonth()] + ", " + new Date(comment.last_updated).getFullYear()
+                  }
+                </div>
+              </p>
+              {comment.authorEmail === email && (<div className='three-dots-menu'>
+                <IconButton
+                  size="very small"
+                  aria-label="more"
+                  id="long-button"
+                  aria-controls={open ? 'long-menu' : undefined}
+                  aria-expanded={open ? 'true' : undefined}
+                  aria-haspopup="true"
+                  onClick={handleClick}
+                >
+                  <MoreVertIcon />
+                </IconButton>
+                <Menu
+                  id="long-menu"
+                  MenuListProps={{
+                    'aria-labelledby': 'long-button',
+                  }}
+                  anchorEl={anchorEl}
+                  open={open}
+                  onClose={handleClose}
+                  PaperProps={{
+                    style: {
+                      maxHeight: ITEM_HEIGHT * 4.5,
+                      width: '20ch',
+                    },
+                  }}
+                >
+                  <MenuItem onClick={handleEdit}>
+                    Edit
+                  </MenuItem>
+                  <MenuItem onClick={handleDelete}>
+                    Delete
+                  </MenuItem>
+                </Menu>
+              </div>
+              )}
+            </div>
+            {editMode ?
+              (
+                <>
+                  <Editor
+                    editorStyle={{ border: "1px solid #C0C0C0" }}
+                    editorState={editCommentState}
+                    onEditorStateChange={setEditCommentState}
+                  />
+                </>
+              ) : (
+                <>
+                  <div>{parse(comment.name)}</div>
+                </>
+              )
+            }
 
             <div style={{ display: "flex", marginTop: "5px" }}>
               {editMode ? (
@@ -133,7 +246,7 @@ const Comment = ({
                                 </>
                               ) : (
                                 <>
-                                  <Like width="15px" height="15px" />
+                                  <Like className="like" />
                                 </>
                               )
                             }
@@ -143,6 +256,10 @@ const Comment = ({
                       </>
                     }
                     handleClick={() => {
+                      if (user === "") {
+                        alert("Please Login to like")
+                        return;
+                      }
                       handleLikeNode(comment.id);
                     }}
                   />
@@ -159,7 +276,7 @@ const Comment = ({
                                 </>
                               ) : (
                                 <>
-                                  <Dislike width="15px" height="15px" />
+                                  <Dislike className="dislike" />
                                 </>
                               )
                             }
@@ -169,6 +286,10 @@ const Comment = ({
                       </>
                     }
                     handleClick={() => {
+                      if (user === "") {
+                        alert("Please Login to dislike")
+                        return;
+                      }
                       handleDislikeNode(comment.id);
                     }}
                   />
@@ -176,57 +297,62 @@ const Comment = ({
                     className="reply"
                     type={
                       <>
-                        {expand ? (
+                        {/* {expand ? (
                           <UpArrow width="10px" height="10px" />
                         ) : (
                           <DownArrow width="10px" height="10px" />
-                        )}{" "}
+                        )}{" "} */}
                         REPLY
                       </>
                     }
                     handleClick={handleNewComment}
                   />
-                  {comment.authorEmail === email && (
-                    <>
-                      <Action
-                        className="reply"
-                        type="EDIT"
-                        handleClick={() => {
-                          setEditMode(true);
-                        }}
-                      />
-                      <Action
-                        className="reply"
-                        type="DELETE"
-                        handleClick={handleDelete}
-                      />
-                    </>
-                  )}
                 </>
               )}
             </div>
           </>
         )}
       </div>
+      {(comment.id !== 1 && comment.items.length > 0) &&
+        <div style={{ display: "flex" }}>
+          <Action
+            className="reply comment"
+            type={
+              <>
+                {expand ? (
+                  <UpArrow width="10px" height="10px" fill='white' />
+                ) : (
+                  <DownArrow width="10px" height="10px" fill='white' />
+                )}{" "}
+                {comment.items.length}
+                {comment.items.length > 1 ? " Replies" : " Reply"}
+              </>
+            }
+            handleClick={handleShowReplies}
+          />
+        </div>
+      }
 
-      <div style={{ display: expand ? "block" : "none", paddingLeft: comment.id === 1 ? 0 : 25 }}>
+      <div style={{ display: expand | showInput ? "block" : "none", paddingLeft: comment.id === 1 ? 0 : 25 }}>
         {showInput && (
-          <div className="inputContainer">
-            <textarea
-              type="text"
-              className="inputContainer__input"
-              autoFocus
-              onChange={(e) => setInput(e.target.value)}
+          <div className="inputContainer reply_textbox">
+            <Editor
+              editorStyle={{ border: "1px solid #C0C0C0" }}
+              editorState={editorState}
+              onEditorStateChange={setEditorState}
+              placeholder="Type your comment here..."
             />
-            <Action className="reply" type="REPLY" handleClick={onAddComment} />
-            <Action
-              className="reply"
-              type="CANCEL"
-              handleClick={() => {
-                setShowInput(false);
-                if (!comment?.items?.length) setExpand(false);
-              }}
-            />
+            <div style={{ display: "flex", flexdirection: "column", marginTop: "5px" }}>
+              <Action className="reply" type="REPLY" handleClick={onAddComment} />
+              <Action
+                className="reply"
+                type="CANCEL"
+                handleClick={() => {
+                  setShowInput(false);
+                  if (!comment?.items?.length) setExpand(false);
+                }}
+              />
+            </div>
           </div>
         )}
